♻️ Refactor tests
